+++ src/entity/npc.h
@@ -24,5 +24,5 @@ class NPC : public Unit
 		
 	public:
 		virtual void update(int delta);
-		virtual int takeDamage(int damage);
+		virtual int takeDamage(float damage);
 };
